Correct silhouette_candidates.py and shortest_path.cpp
When no edge candidates are constructed (i.e. `step` was too
large), then this case was incorrectly handled in both the
Python code that generates the information and the C++ code
which determines the number of edge candidates. The problem is that
a 2d empty array has dimensions (1,0) from NumPy.